refactor: move email verification components to ./app/components
